+++ CHANGES.rst
@@ -4,4 +4,4 @@ Changes
 0.1.0 (2023-02-05)
 ------------------
 * Initial release.
-* Added support for LCG, Basic NCG(FR, PRP, HS, CD, DY, LS, HZ, HS_DY) and two line search function(Armijo & Strong Wolfe).
+* Added support for Linear Conjugate Gradient(LCG), Basic Nonlinear Conjugate Gradient(FR, PRP, HS, CD, DY, LS, HZ, HS-DY) methods and two line search function(Armijo & Strong Wolfe).