+++ CHANGES.rst
@@ -4,4 +4,9 @@ Changes
 0.1.0 (2023-02-05)
 ------------------
 * Initial release.
-* Added support for Linear Conjugate Gradient(LCG), Basic Nonlinear Conjugate Gradient(FR, PRP, HS, CD, DY, LS, HZ, HS-DY) methods and two line search function(Armijo & Strong Wolfe).
+* Added support for Linear Conjugate Gradient (LCG), Basic Nonlinear Conjugate Gradient (i.e. FR, PRP, HS, CD, DY, LS, HZ, HS-DY) methods and two Line Search methods (i.e. Armijo and Strong Wolfe Conditions).
+
+0.2.0 (2024-1-05)
+------------------
+* Second release.
+* Added support for Line Search methods (i.e. Curvature, Weak Wolfe and Goldstein Conditions).